refactor: colocate provider cookie import priorities
