+++ Sources/CodexBarCore/Providers/Copilot/CopilotProviderDescriptor.swift
@@ -1,4 +1,5 @@
 import Foundation
+import SweetCookieKit
 
 public enum CopilotProviderDescriptor {
     public static let descriptor: ProviderDescriptor = Self.makeDescriptor()
@@ -12,6 +13,15 @@ public enum CopilotProviderDescriptor {
             injection: .environment(key: "COPILOT_API_TOKEN"),
             requiresManualCookieSource: false,
             cookieName: nil))
+
+    /// Budget imports stay Chrome-only to avoid prompting unrelated browsers.
+    private static var browserCookieOrder: BrowserCookieImportOrder? {
+        #if os(macOS)
+        [.chrome]
+        #else
+        nil
+        #endif
+    }
 
     static func makeDescriptor() -> ProviderDescriptor {
         ProviderDescriptor(
@@ -36,7 +46,7 @@ public enum CopilotProviderDescriptor {
                 defaultEnabled: false,
                 isPrimaryProvider: false,
                 usesAccountFallback: false,
-                browserCookieOrder: ProviderBrowserCookieDefaults.copilotCookieImportOrder,
+                browserCookieOrder: self.browserCookieOrder,
                 dashboardURL: "https://github.com/settings/copilot",
                 statusPageURL: "https://www.githubstatus.com/"),
             branding: ProviderBranding(
