+++ Sources/CodexBarCore/Providers/LongCat/LongCatProviderDescriptor.swift
@@ -1,10 +1,20 @@
 import Foundation
+import SweetCookieKit
 
 public enum LongCatProviderDescriptor {
     public static let descriptor: ProviderDescriptor = Self.makeDescriptor()
     private static let credentials = ProviderCredentialAdapter(environmentProjections: [
         .cookieHeader(LongCatSettingsReader.cookieHeaderKey, onlyWhenManual: true),
     ])
+
+    /// Preserve Chrome-first behavior, then check Firefox without adding another Keychain prompt.
+    private static var browserCookieOrder: BrowserCookieImportOrder? {
+        #if os(macOS)
+        [.chrome, .firefox]
+        #else
+        nil
+        #endif
+    }
 
     static func makeDescriptor() -> ProviderDescriptor {
         ProviderDescriptor(
@@ -26,7 +36,7 @@ public enum LongCatProviderDescriptor {
                 widgetSelectable: false,
                 isPrimaryProvider: false,
                 usesAccountFallback: false,
-                browserCookieOrder: ProviderBrowserCookieDefaults.longcatCookieImportOrder,
+                browserCookieOrder: self.browserCookieOrder,
                 dashboardURL: "https://longcat.chat/platform/",
                 statusPageURL: nil),
             branding: ProviderBranding(
