+++ Sources/CodexBarCore/Providers/MiMo/MiMoProviderDescriptor.swift
@@ -1,7 +1,17 @@
 import Foundation
+import SweetCookieKit
 
 public enum MiMoProviderDescriptor {
     public static let descriptor: ProviderDescriptor = Self.makeDescriptor()
+
+    /// Safari first, the existing Chrome family, then Firefox/Edge; other Chromium forks remain manual-only.
+    private static var browserCookieOrder: BrowserCookieImportOrder? {
+        #if os(macOS)
+        [.safari, .chrome, .chromeBeta, .chromeCanary, .firefox, .edge]
+        #else
+        nil
+        #endif
+    }
 
     static func makeDescriptor() -> ProviderDescriptor {
         ProviderDescriptor(
@@ -23,7 +33,7 @@ public enum MiMoProviderDescriptor {
                 widgetSelectable: false,
                 isPrimaryProvider: false,
                 usesAccountFallback: false,
-                browserCookieOrder: ProviderBrowserCookieDefaults.mimoCookieImportOrder,
+                browserCookieOrder: self.browserCookieOrder,
                 dashboardURL: "https://platform.xiaomimimo.com/#/console/balance",
                 statusPageURL: nil),
             branding: ProviderBranding(
@@ -128,7 +138,9 @@ struct MiMoWebFetchStrategy: ProviderFetchStrategy {
 
         let sessions = try MiMoCookieImporter.importSessions(browserDetection: context.browserDetection)
         guard !sessions.isEmpty else {
-            if let lastError { throw lastError }
+            if let lastError {
+                throw lastError
+            }
             throw MiMoSettingsError.missingCookie()
         }
 
@@ -151,7 +163,9 @@ struct MiMoWebFetchStrategy: ProviderFetchStrategy {
             }
         }
 
-        if let lastError { throw lastError }
+        if let lastError {
+            throw lastError
+        }
         throw MiMoSettingsError.missingCookie()
         #else
         throw MiMoSettingsError.missingCookie()
