+++ Sources/CodexBarCore/Providers/Providers.swift
@@ -199,120 +199,9 @@ public enum ProviderDefaults {
 }
 
 public enum ProviderBrowserCookieDefaults {
-    public static var chromeOnlyImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.chrome]
-        #else
-        nil
-        #endif
-    }
-
     public static var defaultImportOrder: BrowserCookieImportOrder? {
         #if os(macOS)
         Browser.defaultImportOrder
-        #else
-        nil
-        #endif
-    }
-
-    /// Safari first for Cursor: active sessions often live only there, and Chromium profiles may carry stale tokens.
-    public static var cursorCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.safari] + Browser.defaultImportOrder.filter { $0 != .safari }
-        #else
-        nil
-        #endif
-    }
-
-    /// Preserve the legacy Codex prompt behavior: prefer Safari/Chrome/Firefox before
-    /// probing additional Chromium variants that may trigger Safe Storage prompts.
-    public static var codexCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        let preferredPrefix: [Browser] = [.safari, .chrome, .firefox]
-        return preferredPrefix + Browser.defaultImportOrder.filter { !preferredPrefix.contains($0) }
-        #else
-        nil
-        #endif
-    }
-
-    /// OpenCode web Auto stays Chrome-only by default, with Dia as the one bounded provider exception
-    /// because Dia has a confirmed reporter need. Other browsers stay on Manual until users can choose them.
-    public static var opencodeCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.chrome, .dia]
-        #else
-        nil
-        #endif
-    }
-
-    /// Grok is normally signed in through Chrome; keep this narrow so CLI/live probes do not touch
-    /// unrelated browser keychains.
-    public static var grokCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.chrome]
-        #else
-        nil
-        #endif
-    }
-
-    /// MiMo Auto: Safari first (no Keychain prompt), keep the existing Chrome-family
-    /// entries from main, and add Firefox/Edge per #1304. Other Chromium forks stay on
-    /// Manual import to avoid scanning the full SweetCookieKit default order.
-    public static var mimoCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.safari, .chrome, .chromeBeta, .chromeCanary, .firefox, .edge]
-        #else
-        nil
-        #endif
-    }
-
-    /// Devin sessions are normally in Chrome. Keep automatic import narrow so live probes do not
-    /// touch unrelated browser keychains; users can select another browser explicitly.
-    public static var devinCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.chrome]
-        #else
-        nil
-        #endif
-    }
-
-    /// Copilot budget imports should stay Chrome-only by default to avoid prompting unrelated browsers.
-    public static var copilotCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.chrome]
-        #else
-        nil
-        #endif
-    }
-
-    /// LongCat Auto keeps Chrome first for existing users, then checks Firefox without adding
-    /// an unrelated browser Keychain prompt.
-    public static var longcatCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.chrome, .firefox]
-        #else
-        nil
-        #endif
-    }
-
-    /// Qoder sessions are documented through Chrome cookie import. Keep automatic import narrow
-    /// so enabling this provider does not probe unrelated browser keychains.
-    public static var qoderCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.chrome]
-        #else
-        nil
-        #endif
-    }
-
-    /// Mistral Auto: Chrome first (matches the original Chrome-only behavior so
-    /// existing users see no change), then Firefox so users signed in via Firefox
-    /// or Firefox Developer Edition are detected without Manual mode. Safari
-    /// follows for Full Disk Access users. Other Chromium forks stay on Manual
-    /// import to avoid scanning the full default order.
-    public static var mistralCookieImportOrder: BrowserCookieImportOrder? {
-        #if os(macOS)
-        [.chrome, .firefox, .safari]
         #else
         nil
         #endif
