+++ Tests/CodexBarTests/BrowserCookieOrderLabelTests.swift
@@ -52,7 +52,6 @@ struct BrowserCookieOrderStatusStringTests {
     func `opencode automatic cookies only use chrome and dia`() {
         let order = OpenCodeWebCookieSupport.automaticImportOrder(provider: .opencode)
         #expect(order == ProviderDefaults.metadata[.opencode]?.browserCookieOrder)
-        #expect(order == ProviderBrowserCookieDefaults.opencodeCookieImportOrder)
         #expect(order == [.chrome, .dia])
     }
 
@@ -73,7 +72,6 @@ struct BrowserCookieOrderStatusStringTests {
     @Test
     func `mimo cookie import order supports safari firefox and edge`() {
         let order = ProviderDefaults.metadata[.mimo]?.browserCookieOrder ?? Browser.defaultImportOrder
-        #expect(order == ProviderBrowserCookieDefaults.mimoCookieImportOrder)
         #expect(order == [.safari, .chrome, .chromeBeta, .chromeCanary, .firefox, .edge])
         #expect(order.first == .safari)
         #expect(order.contains(.firefox))
@@ -84,13 +82,11 @@ struct BrowserCookieOrderStatusStringTests {
     @Test
     func `copilot cookie imports default to chrome only`() {
         #expect(ProviderDefaults.metadata[.copilot]?.browserCookieOrder == [.chrome])
-        #expect(ProviderBrowserCookieDefaults.copilotCookieImportOrder == [.chrome])
     }
 
     @Test
     func `mistral cookie import order supports chrome firefox and safari`() {
         let order = ProviderDefaults.metadata[.mistral]?.browserCookieOrder ?? Browser.defaultImportOrder
-        #expect(order == ProviderBrowserCookieDefaults.mistralCookieImportOrder)
         #expect(order == [.chrome, .firefox, .safari])
         #expect(order.first == .chrome)
         #expect(order.contains(.firefox))
@@ -104,13 +100,11 @@ struct BrowserCookieOrderStatusStringTests {
     @Test
     func `longcat cookie import order supports chrome and firefox`() {
         let metadataOrder = ProviderDefaults.metadata[.longcat]?.browserCookieOrder
-        let defaultOrder = ProviderBrowserCookieDefaults.longcatCookieImportOrder
 
         #expect(metadataOrder == [.chrome, .firefox])
-        #expect(defaultOrder == [.chrome, .firefox])
-        #expect(defaultOrder?.first == .chrome)
-        #expect(defaultOrder?.contains(.firefox) == true)
-        #expect(defaultOrder?.contains(.safari) == false)
+        #expect(metadataOrder?.first == .chrome)
+        #expect(metadataOrder?.contains(.firefox) == true)
+        #expect(metadataOrder?.contains(.safari) == false)
     }
     #endif
 }
